+++ src/python/pants/util/BUILD
@@ -5,9 +5,29 @@ python_tests(
   name='tests',
   dependencies=[
     '3rdparty/python:dataclasses',
+    'src/python/pants/testutil:test_base',
+    ':argutil',
+    ':collections',
+    ':contextutil',
+    ':dirutil',
+    ':enums',
+    ':eval',
+    ':fileutil',
     ':filtering',
     ':frozendict',
+    ':memo',
+    ':meta',
+    ':netrc',
+    ':objects',
     ':ordered_set',
+    ':osutil',
+    ':process_handler',
+    ':retry',
+    ':rwbuf',
+    ':socket',
+    ':strutil',
+    ':tarutil',
+    ':xml_parser',
   ],
   tags = {'type_checked'},
 )
