+++ src/python/pants/util/collections_test.py
