+++ src/python/pants/util/memo_test.py
