+++ src/python/pants/util/objects_test.py
